fix __dirname is not defined
