+++ ui/desktop/src/components/settings/keyboard/ShortcutRecorder.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useRef } from 'react';
 import { Button } from '../../ui/button';
-import { KeyboardShortcuts } from '../../../utils/settings';
+import { KeyboardShortcuts } from '../../../utils/settingsTypes';
 import { getShortcutLabel, formatShortcut } from './KeyboardShortcutsSection';
 
 interface ShortcutRecorderProps {
