+++ ui/desktop/src/preload.ts
@@ -1,7 +1,7 @@
 import Electron, { contextBridge, ipcRenderer, webUtils } from 'electron';
 import { Recipe } from './recipe';
 import { GooseApp } from './api';
-import type { Settings } from './utils/settings';
+import type { Settings } from './utils/settingsTypes';
 
 interface NotificationData {
   title: string;
